Added ResponseError and utility to error wrap in Singles

+++ src/main/java/com/microsoft/azure/storage/blob/AppendBlobAccessConditions.java
@@ -31,9 +31,9 @@ public final class AppendBlobAccessConditions {
     public static final AppendBlobAccessConditions NONE =
             new AppendBlobAccessConditions(null, null);
 
-    private final Integer ifAppendPositionEquals;
+    private final Long ifAppendPositionEquals;
 
-    private final Integer ifMaxSizeLessThanOrEqual;
+    private final Long ifMaxSizeLessThanOrEqual;
 
     /**
      * Creates a {@code AppendBlobAccessConditions} object.
@@ -44,7 +44,7 @@ public final class AppendBlobAccessConditions {
      *      Ensures that the AppendBlock operation succeeds only if the append blob's size is less than or
      *      equal to a value.
      */
-    public AppendBlobAccessConditions(Integer ifAppendPositionEquals, Integer ifMaxSizeLessThanOrEqual) {
+    public AppendBlobAccessConditions(Long ifAppendPositionEquals, Long ifMaxSizeLessThanOrEqual) {
         if ((ifAppendPositionEquals != null && ifAppendPositionEquals < -1) ||
                 (ifMaxSizeLessThanOrEqual != null && ifMaxSizeLessThanOrEqual < -1)) {
             throw new IllegalArgumentException("Append blob access conditions can't be less than -1.");
@@ -57,7 +57,7 @@ public final class AppendBlobAccessConditions {
      * @return
      *      Ensures that the AppendBlock operation succeeds only if the append position is equal to the value.
      */
-    public Integer getIfAppendPositionEquals() {
+    public Long getIfAppendPositionEquals() {
         return ifAppendPositionEquals;
     }
 
@@ -66,7 +66,7 @@ public final class AppendBlobAccessConditions {
      *      Ensures that the AppendBlock operation succeeds only if the append blob's size is less than or equal to the
      *      value.
      */
-    public Integer getIfMaxSizeLessThanOrEqual() {
+    public Long getIfMaxSizeLessThanOrEqual() {
         return ifMaxSizeLessThanOrEqual;
     }
 }
